Add AdministrationOnlyMixin for authorization in Maintenance app views
This commit introduces the AdministrationOnlyMixin to the Maintenance application. This mixin is used to add an authorization layer to the class-based views in the Maintenance app, ensuring that only users with an ADMINISTRATION user type can access certain views. Non-admin users are redirected to the home page. This enhances the security and access control of the Maintenance application.